Refactored trackers and skill handling

+++ app/src/main/java/com/infonuascape/osrshelper/hiscore/HiscoreFetcher.java
@@ -61,7 +61,7 @@ public class HiscoreFetcher {
 			dataSeparator = skillArray[i].split(",");
 			skillList.get(i).setRank(Integer.parseInt(dataSeparator[0]));
 			skillList.get(i).setLevel(Short.parseShort(dataSeparator[1]));
-			skillList.get(i).setExperience(Long.parseLong(dataSeparator[2]));
+			skillList.get(i).setExperience(Integer.parseInt(dataSeparator[2]));
 		}
 		return ps;
 	}
